UI for todos and initial setup for backend Fixes #5
