+++ src/app/api/notes/route.js
@@ -1,0 +1,1 @@
+// NOTES API
