+++ src/app/todos/page.jsx
@@ -1,5 +1,5 @@
 import PageView from "@/components/SharedComponents/PageView";
-import TodoWindow from "@/components/SharedComponents/Todos/TodoWindow";
+import TodoWindow from "@/components/Todos/TodoWindow";
 
 function Todospage() {
   return <PageView window={<TodoWindow />}></PageView>;
