+++ src/components/Calculator/CalculatorHeader.jsx
@@ -1,0 +1,8 @@
+import WindowHeader from "../SharedComponents/WindowHeader";
+import { CALCULATOR } from "@/helpers/Constants";
+
+function CalculatorHeader() {
+  return <WindowHeader windowName={CALCULATOR}></WindowHeader>;
+}
+
+export default CalculatorHeader;
