+++ src/components/SharedComponents/Hero.jsx
@@ -1,10 +1,10 @@
 function Hero() {
   return (
     <section className="flex flex-col py-4 px-4 gap-4">
-      <div className="text-white text-center px-2 py-2 m-4">
+      <div className="text-center px-2 py-2 m-4">
         <h1 className="text-6xl font-mono">Your Productivity Hub</h1>
       </div>
-      <div className="text-white text-center px-2 py-2 m-4">
+      <div className="text-center px-2 py-2 m-4">
         <p className="text-lg font-mono">
           All your tools in one place for maximum efficiency
         </p>
