+++ src/components/Todos/TodosHeader.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import WindowHeader from "../SharedComponents/WindowHeader";
+import { TODOS } from "@/helpers/Constants";
+
+function TodosHeader() {
+  return <WindowHeader windowName={TODOS}></WindowHeader>;
+}
+
+export default TodosHeader;
